Fix Versioned Resource tests, add placeholder TimeMap type
Revert HttpResourceConverter changes
