+++ fcrepo-http-api/src/test/java/org/fcrepo/integration/http/api/AbstractResourceIT.java
@@ -44,9 +44,10 @@ import static org.slf4j.LoggerFactory.getLogger;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
+import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Collection;
-
+import javax.ws.rs.core.Link;
 import javax.ws.rs.core.Response.Status;
 import javax.xml.bind.DatatypeConverter;
 
@@ -498,6 +499,22 @@ public abstract class AbstractResourceIT {
         }
         ttl.append(" <" + predicateUri + "> ");
         ttl.append(literal);
+    }
+
+    /**
+     * Test a response for a specific LINK header
+     * 
+     * @param response the HTTP response
+     * @param uri the URI expected in the LINK header
+     * @param rel the rel argument to check for
+     */
+    protected void checkForLinkHeader(final CloseableHttpResponse response, final String uri, final String rel) {
+        final Link linkA = Link.valueOf("<" + uri + ">; rel=" + rel);
+        final int count = (int) Arrays.asList(response.getHeaders(LINK)).stream().filter(x -> {
+            final Link linkB = Link.valueOf(x.getValue());
+            return linkB.equals(linkA);
+        }).count();
+        assertEquals(1, count);
     }
 
 }
