+++ fcrepo-http-commons/src/main/java/org/fcrepo/http/commons/api/rdf/HttpResourceConverter.java
@@ -53,7 +53,6 @@ import org.fcrepo.kernel.api.exception.TombstoneException;
 import org.fcrepo.kernel.api.identifiers.IdentifierConverter;
 import org.fcrepo.kernel.api.models.NonRdfSourceDescription;
 import org.fcrepo.kernel.api.models.FedoraResource;
-import org.fcrepo.kernel.api.models.FedoraTimeMap;
 import org.fcrepo.kernel.modeshape.TombstoneImpl;
 import org.fcrepo.kernel.modeshape.identifiers.HashConverter;
 import org.fcrepo.kernel.modeshape.identifiers.NamespaceConverter;
@@ -115,18 +114,12 @@ public class HttpResourceConverter extends IdentifierConverter<Resource,FedoraRe
         final Session jcrSession = getJcrSession(session);
         try {
             if (path != null) {
-
                 final Node node = getNode(path);
 
                 final boolean metadata = values.containsKey("path")
                         && values.get("path").endsWith("/" + FCR_METADATA);
-                final boolean timemap = values.containsKey("path") && values.get("path").endsWith("/" + FCR_VERSIONS);
 
                 final FedoraResource fedoraResource = nodeConverter.convert(node);
-
-                if (timemap) {
-                    return fedoraResource.findOrCreateTimeMap();
-                }
 
                 if (!metadata && fedoraResource instanceof NonRdfSourceDescription) {
                     return fedoraResource.getDescribedResource();
@@ -210,14 +203,10 @@ public class HttpResourceConverter extends IdentifierConverter<Resource,FedoraRe
         if (uriTemplate.match(resource.getURI(), values) && values.containsKey("path")) {
             String path = "/" + values.get("path");
 
-            final boolean timemap = path.endsWith("/" + FCR_VERSIONS);
             final boolean metadata = path.endsWith("/" + FCR_METADATA);
 
             if (metadata) {
                 path = replaceOnce(path, "/" + FCR_METADATA, EMPTY);
-            }
-            if (timemap) {
-                path = replaceOnce(path, "/" + FCR_VERSIONS, EMPTY);
             }
 
             path = forward.convert(path);
@@ -260,13 +249,31 @@ public class HttpResourceConverter extends IdentifierConverter<Resource,FedoraRe
     }
 
     private static String getPath(final FedoraResource resource) {
-        if (resource instanceof FedoraTimeMap) {
-            // the path is relative to the parent, and unique so fake it here.
-            final FedoraResource parent = resource.getContainer();
+        if (resource.isFrozenResource()) {
+            // the versioned resource we're in
+            final FedoraResource versionableFrozenResource = resource.getVersionedAncestor();
 
-            final String parentPath = parent.getPath();
+            // the unfrozen equivalent for the versioned resource
+            final FedoraResource unfrozenVersionableResource = versionableFrozenResource.getUnfrozenResource();
 
-            return parentPath + "/" + FCR_VERSIONS;
+            // the label for this version
+            final String versionLabel = versionableFrozenResource.getVersionLabelOfFrozenResource();
+
+            // the path to this resource within the versioning tree
+            final String pathWithinVersionable;
+
+            if (!resource.equals(versionableFrozenResource)) {
+                pathWithinVersionable = getRelativePath(resource, versionableFrozenResource);
+            } else {
+                pathWithinVersionable = "";
+            }
+
+            // and, finally, the path we want to expose in the URI
+            final String path = unfrozenVersionableResource.getPath()
+                    + "/" + FCR_VERSIONS
+                    + (versionLabel != null ? "/" + versionLabel : "")
+                    + pathWithinVersionable;
+            return path.startsWith("/") ? path : "/" + path;
         }
         return resource.getPath();
     }
